Implemented create, delete and modify comment, and displays them on bottom. Still requires user specific comments

+++ website/review.py
@@ -2,25 +2,29 @@
 from flask import Blueprint, request, flash, jsonify
 import json
 
-from .models import Comments
+from .models import Comments, Recipes
 from . import db
 
 review = Blueprint('review', __name__)
 
 # Creates a comment given the recipe_id
 # Already checks http request 
-def create_comment(recipe_id):
-    comment = request.form.get('review')
-    print(recipe_id)
-    print(comment)
+@review.route('/create-comment', methods=['POST'])
+def create_comment():
+    print("hewahew")
+    if request.method == 'POST':
+        recipe = json.loads(request.data)
+        print(recipe)
+        recipe_id = recipe['recipe_id']
+        comment = recipe['comment']
 
-    if len(comment) < 1:
-        flash('Comment is too short!', category='error')
-    else:
-        new_comment = Comments(comment=comment, has=recipe_id)
-        db.session.add(new_comment)
-        db.session.commit()
-        flash('Comment added!', category='success')
+        if len(comment) < 1:
+            flash('Comment is too short!', category='error')
+        else:
+            new_comment = Comments(comment=comment, has=recipe_id)
+            db.session.add(new_comment)
+            db.session.commit()
+            flash('Comment added!', category='success')
 
 # Retrieves comments given recipe_id
 def retrieve_comments(recipe_id):
@@ -28,11 +32,29 @@ def retrieve_comments(recipe_id):
 
 @review.route('/delete-comment', methods=['POST'])
 def delete_comment():
-    comment = json.loads(request.data)
-    comment_id = comment['comment_id']
-    comment = Comments.query.get(comment_id)
-    print(comment)
-    db.session.delete(comment)
-    db.session.commit()
+    if request.method == 'POST':
+        comment = json.loads(request.data)
+        comment_id = comment['comment_id']
+        comment = Comments.query.get(comment_id)
+        print(comment)
 
-    return jsonify({})
+
+        db.session.delete(comment)
+        db.session.commit()
+        flash('Deleted comment successfully!', category='success')
+
+        return jsonify({})
+
+@review.route('/modify-comment', methods=['POST'])
+def modify_comment():
+    if request.method == 'POST':
+        comment = json.loads(request.data)
+        comment_id = comment['comment_id']
+        new_comment = comment['comment']
+
+        comment = Comments.query.get(comment_id)
+        comment.comment = new_comment
+        db.session.commit()
+        flash('Modified comment successfully!', category='success')
+
+        return jsonify({})
